+++ Branch-SDK/src/io/branch/referral/Defines.java
@@ -14,6 +14,7 @@ public class Defines {
         DeviceFingerprintID("device_fingerprint_id"),
         SessionID("session_id"),
         LinkClickID("link_click_id"),
+        GoogleSearchInstallReferrer("google_search_install_referrer"),
         FaceBookAppLinkChecked("facebook_app_link_checked"),
         BranchLinkUsed("branch_used"),
         ReferringBranchIdentity("referring_branch_identity"),
@@ -31,6 +32,7 @@ public class Defines {
         Expiration("expiration"),
         Event("event"),
         Metadata("metadata"),
+        CommerceData("commerce_data"),
         ReferralCode("referral_code"),
         Total("total"),
         Unique("unique"),
@@ -114,7 +116,7 @@ public class Defines {
         ContentAnalyticsMode("content_analytics_mode"),
         ContentDiscovery("cd"),
         InstallReferrer("install_referrer");
-        
+
         private String key = "";
 
         Jsonkey(String key) {