+++ Branch-SDK/src/io/branch/referral/util/ShareSheetStyle.java
@@ -11,6 +11,8 @@ import android.support.annotation.StyleRes;
 import android.view.View;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 import io.branch.indexing.BranchUniversalObject;
 import io.branch.referral.Branch;
@@ -44,6 +46,9 @@ public class ShareSheetStyle {
 
     private String sharingTitle_ = null;
     private View sharingTitleView_ = null;
+
+    private List<String> includeInShareSheet = new ArrayList<>();
+    private List<String> excludeFromShareSheet = new ArrayList<>();
 
     public ShareSheetStyle(@NonNull Context context, @NonNull String messageTitle, @NonNull String messageBody) {
         context_ = context;
@@ -200,6 +205,86 @@ public class ShareSheetStyle {
     public ShareSheetStyle setSharingTitle(View titleView) {
         this.sharingTitleView_ = titleView;
         return this;
+    }
+
+    /**
+     * Exclude items from the ShareSheet by package name String.
+     *
+     * @param packageName {@link String} package name to be excluded.
+     * @return this Builder object to allow for chaining of calls to set methods.
+     */
+    public ShareSheetStyle excludeFromShareSheet(@NonNull String packageName) {
+        this.excludeFromShareSheet.add(packageName);
+        return this;
+    }
+
+    /**
+     * Exclude items from the ShareSheet by package name array.
+     *
+     * @param packageName {@link String[]} package name to be excluded.
+     * @return this Builder object to allow for chaining of calls to set methods.
+     */
+    public ShareSheetStyle excludeFromShareSheet(@NonNull String[] packageName) {
+        excludeFromShareSheet.addAll(Arrays.asList(packageName));
+        return this;
+    }
+
+    /**
+     * Exclude items from the ShareSheet by package name List.
+     *
+     * @param packageNames {@link List} package name to be excluded.
+     * @return this Builder object to allow for chaining of calls to set methods.
+     */
+    public ShareSheetStyle excludeFromShareSheet(@NonNull List<String> packageNames) {
+        this.excludeFromShareSheet.addAll(packageNames);
+        return this;
+    }
+
+    /**
+     * Include items from the ShareSheet by package name String. If only "com.Slack"
+     * is included, then only preferred sharing options + Slack
+     * will be displayed, for example.
+     *
+     * @param packageName {@link String} package name to be included.
+     * @return this Builder object to allow for chaining of calls to set methods.
+     */
+    public ShareSheetStyle includeInShareSheet(@NonNull String packageName) {
+        this.includeInShareSheet.add(packageName);
+        return this;
+    }
+
+    /**
+     * Include items from the ShareSheet by package name Array. If only "com.Slack"
+     * is included, then only preferred sharing options + Slack
+     * will be displayed, for example.
+     *
+     * @param packageName {@link String[]} package name to be included.
+     * @return this Builder object to allow for chaining of calls to set methods.
+     */
+    public ShareSheetStyle includeInShareSheet(@NonNull String[] packageName) {
+        includeInShareSheet.addAll(Arrays.asList(packageName));
+        return this;
+    }
+
+    /**
+     * Include items from the ShareSheet by package name List. If only "com.Slack"
+     * is included, then only preferred sharing options + Slack
+     * will be displayed, for example.
+     *
+     * @param packageNames {@link List} package name to be included.
+     * @return this Builder object to allow for chaining of calls to set methods.
+     */
+    public ShareSheetStyle includeInShareSheet(@NonNull List<String> packageNames) {
+        this.includeInShareSheet.addAll(packageNames);
+        return this;
+    }
+
+    public List<String> getExcludedFromShareSheet() {
+        return excludeFromShareSheet;
+    }
+
+    public List<String> getIncludedInShareSheet() {
+        return includeInShareSheet;
     }
 
     public ArrayList<SharingHelper.SHARE_WITH> getPreferredOptions() {